+++ ProgramSubjectCrawlerService/Crawlers/StudentProgramCrawler.cs
@@ -242,7 +242,7 @@ namespace ProgramSubjectCrawlerService.Crawlers
 
                 //study unit
                 HtmlNode tdStudyUnitTag = docNode.SelectSingleNode("//td[3]");
-                int studyUnit = int.Parse(StringHelper.SuperCleanString(tdStudyUnitTag.InnerText));
+                byte studyUnit = byte.Parse(StringHelper.SuperCleanString(tdStudyUnitTag.InnerText));
 
                 //studyunit type
                 HtmlNode tdStudyUnitTypeTag = docNode.SelectSingleNode("//td[4]");
