+++ cs4rsa_core/Models/SubjectModel.cs
@@ -17,7 +17,7 @@ namespace cs4rsa_core.Models
         public List<ClassGroupModel> ClassGroupModels { get; set; }
         public string SubjectName => _subject.Name;
         public string SubjectCode => _subject.SubjectCode;
-        public int StudyUnit { get; set; }
+        public byte StudyUnit { get; set; }
         public int CourseId => _subject.CourseId;
         public string StudyUnitType => _subject.StudyUnitType;
         public string StudyType => _subject.StudyType;
